Use UInt for offset in highlight context
Julia CI indicates that on 32-bit machines, a method error occurs when
HighightContext is attempted to be constructed using a UInt32 when only
an Int32 is allowed. Since the only thing that happens with the offset
is addition, and unsigned integers "win" in addition (i.e. the result is
unsigned) I suspect that unsigned-ness spreads from our use of
JuliaSyntax. If this is indeed the case, then simply switching the
offset from an Int to UInt will resolve this issue.