Guard against the assumption children gives nodes
I've recently become aware that some dodgy syntax can result in
numchildren(...) > 0 but no children(...), represented as nothing.

We might as well systematically remove the chance of running into this.

If this isn't the end of the syntax surprises, I'll look at what's
needed to make JET confident there aren't any more lurking around the
corner (this is probably a good idea anyway, but spinning plates and all
that).